Refactor : 헥사고날 패키지 구조 컨벤션에 맞게 수정
