+++ src/main/java/com/promptoven/cartservice/adapter/web/controller/CartRestController.java
@@ -1,12 +1,12 @@
-package com.promptoven.cartservice.adapter.in.web.controller;
+package com.promptoven.cartservice.adapter.web.controller;
 
-import com.promptoven.cartservice.adapter.in.web.mapper.CartVoMapper;
-import com.promptoven.cartservice.adapter.in.web.vo.CartCreateRequestVo;
-import com.promptoven.cartservice.adapter.in.web.vo.CartDeleteRequestVo;
-import com.promptoven.cartservice.adapter.in.web.vo.CartResponseVo;
-import com.promptoven.cartservice.adapter.in.web.vo.CartUpdateVo;
-import com.promptoven.cartservice.application.port.in.CartRequestDto;
-import com.promptoven.cartservice.application.port.in.CartUseCase;
+import com.promptoven.cartservice.adapter.web.mapper.CartVoMapper;
+import com.promptoven.cartservice.adapter.web.vo.CartCreateRequestVo;
+import com.promptoven.cartservice.adapter.web.vo.CartDeleteRequestVo;
+import com.promptoven.cartservice.adapter.web.vo.CartResponseVo;
+import com.promptoven.cartservice.adapter.web.vo.CartUpdateVo;
+import com.promptoven.cartservice.application.port.dto.in.CartRequestDto;
+import com.promptoven.cartservice.application.port.call.CartUseCase;
 import com.promptoven.cartservice.global.common.response.BaseResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
