+++ src/main/java/com/promptoven/cartservice/adapter/web/vo/CartResponseVo.java
@@ -1,4 +1,4 @@
-package com.promptoven.cartservice.adapter.in.web.vo;
+package com.promptoven.cartservice.adapter.web.vo;
 
 import java.time.LocalDateTime;
