+++ src/main/java/com/promptoven/cartservice/application/port/call/CartUseCase.java
@@ -1,4 +1,6 @@
-package com.promptoven.cartservice.application.port.in;
+package com.promptoven.cartservice.application.port.call;
+
+import com.promptoven.cartservice.application.port.dto.in.CartRequestDto;
 
 import java.util.List;
