+++ src/main/java/com/promptoven/cartservice/domain/service/CartDomainService.java
@@ -4,8 +4,8 @@ import java.util.List;
 
 import org.springframework.stereotype.Service;
 
-import com.promptoven.cartservice.application.port.in.CartRequestDto;
-import com.promptoven.cartservice.application.port.out.CartOutportDto;
+import com.promptoven.cartservice.application.port.dto.in.CartRequestDto;
+import com.promptoven.cartservice.application.port.dto.out.CartOutportDto;
 import com.promptoven.cartservice.domain.model.Cart;
 
 @Service
