+++ src/main/java/com/promptoven/cartservice/global/config/KafkaConsumerConfig.java
@@ -1,6 +1,6 @@
 package com.promptoven.cartservice.global.config;
 
-import com.promptoven.cartservice.application.port.in.RequestMessageDto;
+import com.promptoven.cartservice.application.port.dto.in.RequestMessageDto;
 import org.apache.kafka.clients.consumer.ConsumerConfig;
 import org.apache.kafka.common.serialization.StringDeserializer;
 import org.springframework.beans.factory.annotation.Value;
